apparently postgres doesn't like my tables anymore

+++ backend/db/migrations/20220704173935-create-spot.js
@@ -33,11 +33,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.FLOAT(3, 15),
+        type: Sequelize.DECIMAL,
         allowNull: false
       },
       lng: {
-        type: Sequelize.FLOAT(3, 15),
+        type: Sequelize.DECIMAL,
         allowNull: false
       },
       price: {
